Starting to refactor the views layout, to make it tiled.
The idea is to move the data type from, say, Scalar*[NP][NP][NUM_LEV] to
Scalar*[NUM_LEV][NP][NP]. Notice that this is not the same as the flipped
version of the fortran layout (np,np,nlev,nelemd), since the data type
Scalar still has VECTOR_SIZE doubles in it. This is why we refer to this
layout as tiled.

+++ components/homme/test_execs/share_ut/share_ut.cpp
@@ -35,7 +35,7 @@ void genRandArray(Real *const x, int length, rngAlg &engine, PDF &pdf) {
 template <typename ViewType, typename rngAlg, typename PDF>
 void genRandArray(ViewType view, rngAlg &engine, PDF &pdf) {
   Real *data = view.data();
-  for (int i = 0; i < view.size(); ++i) {
+  for (size_t i = 0; i < view.size(); ++i) {
     data[i] = pdf(engine);
   }
 }
